api/requirements.txt
psycopg2-binary 
fastapi[standard]